removed a few unnecessary casts

+++ src/jd/core/util/StringUtil.java
@@ -44,8 +44,8 @@ public class StringUtil
 			if (c < ' ')
 			{
 				sb.append("\\0");
-				sb.append((char)('0' + ((int)c >> 3)));
-				sb.append((char)('0' + ((int)c & 7)));
+				sb.append((char)('0' + (c >> 3)));
+				sb.append((char)('0' + (c & 7)));
 			}
 			else
 			{
